EEPROM uint16 change
Changed the address type from uint8 to uint16 so we can access all the 4000 something addresses in the EEPROM instead of being locked to 256 addresses with an uint8.

+++ src/EEPROM_promter.c
@@ -1,7 +1,7 @@
 #include <avr/io.h>
 #include <EEPROM_promter.h>
 
-void EEPROM_write(uint8_t uiAddress, uint8_t ucData) {
+void EEPROM_write(uint16_t uiAddress, uint8_t ucData) {
   // Wait for completion of previous write /
   while(EECR & (1<<EEPE))
   ;
@@ -15,7 +15,7 @@ void EEPROM_write(uint8_t uiAddress, uint8_t ucData) {
   EECR |= (1<<EEPE);
 }
 
-uint8_t EEPROM_read(uint8_t uiAddress) {
+uint8_t EEPROM_read(uint16_t uiAddress) {
   // Wait for completion of previous write /
   while(EECR & (1<<EEPE))
   ;
