+++ src/EEPROM_promter.h
@@ -5,9 +5,9 @@
 #include <stdint.h>
 
 // Function to write data to EEPROM /
-void EEPROM_write(uint8_t uiAddress, uint8_t ucData);
+void EEPROM_write(uint16_t uiAddress, uint8_t ucData);
 
 // Function to read data from EEPROM /
-uint8_t EEPROM_read(uint8_t uiAddress);
+uint8_t EEPROM_read(uint16_t uiAddress);
 
 #endif // EEPROM_H
